Support window resizing
These were originally 12 separate commits, but this was quite a hassle
when rebasing with all the frequent breaking changes on the main branch:

Target a baseview branch with window resizing

Add field on ApplicationRunner for new window size

This is needed because we don't own the baseview `window` object, so we
can't interact with it directly from within a view.

Add a WindowDescription field for scale factor

This would be a custom scale factor applied on top of the DPI scaling.
It's useful to keep track of this separately because otherwise dealing
with external DPI changes will be very difficult.

Use the user scale factor in vizia_baseview

This is additional scaling on top of the regular HiDPI scaling. This is
useful to allow GUIs to scale uniformly independently of the system's
HiDPI settings.

Handle resizing and scaling events in baseview

Only perform window resizing on Linux

Since the targetted baseview branch currently only implements it there.

Register the baseview window state model

And move it to a separate reference counted struct to make it happen.

Recompute props when window size or scale changes

Fix artifacts when resizing baseview GUIs

The logical width was used to compute the new physical height.

Print a warning when resizing on a non-Linux OS

Resize and rescale through cx instead of events

Changing a property instead of emitting an event makes the handling much
clearer and it also prevents multiple resizes from happening during a
single frame.

Add a user scale factor modifier

Now that you can no longer pass a `WindowDescription` directly.

Partially rewritten when rebasing on top of 5b0c9a434440eff96e1e96d5410421c2423c3161.

Partially rewritten when rebasing on top of 8299ccf921b09b15de77195a185d5673e96c3200.