Avoid changing logical size when changing scale
This happened due to rounding errors in the conversion to and from
integers. We can detect this by doing the same computation, and treating
the resize as a noop if it does happen. This prevents the logical size
from slowly drifting when changing the user scale many times in a row.